feat: add description voto
